pyproject.toml
[project]
name = "offsec-challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "langchain (>=0.3.26,<0.4.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "ldap3 (>=2.9.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
